Move: MessageGenerator.kt 파일 위치 view 패키지로 이동
