+++ src/test/kotlin/carRacing/MessageGeneratorTest.kt
@@ -1,7 +1,7 @@
 package carRacing
 
-import carRacing.domain.MessageGenerator
-import carRacing.domain.VARIABLE_PATTERN
+import carRacing.view.MessageGenerator
+import carRacing.view.VARIABLE_PATTERN
 import carRacing.enum.ErrorCode
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
